+++ NEWS.md
@@ -4,6 +4,7 @@
 - Added a couple input checks for the `obsID` and `outcome` arguments.
 - `vcov.logitr()` method now returns `object$vcov` if the user set `vcov = TRUE` during estimation (avoids a redundant calculation of `vcov`).
 - Added new datasets: `apolloModeChoiceData`, `electricity`
+- Added support for new mixed logit distributions: zero-censored normal
 
 # logitr 0.5.1